+++ docs/main/changelog.rst
@@ -7,8 +7,8 @@ that were made in every particular version.
 From version 0.7.6 *Dependency Injector* framework strictly 
 follows `Semantic versioning`_
 
-Develop
--------
+4.4.1
+-----
 - Improve ``FastAPI`` integration: handle ``Depends(Provide[...])``.
 - Update ``FastAPI`` example.
 - Remove a typo from the ``Flask`` tutorial.